--- storage/mock.go
@@ -1,6 +1,0 @@
-package storage
-
-
-//TODO: Mocking database for testing
-//Ideally this should be something memory based
-
